Make LazyDatabase
LazyDatabase essentially copies some of the functionality of
LevenshteinGraph while adapting it to inherit LevenshteinDatabase. The
purpose of doing this is to eventually allow the Levenshtein algorithms
to pick and choose different database structures by decoupling the
dictionary and neighbor searching (the job of the database) with the job
of the graph.