Refactor wildcard identity getting and putting
Instead of having to first get every identity and iterate over that list
to add the identities of a wildcard, wildcard identities can be added
while being found without having to put them into a secondary list. This
has been enabled by adding a functional interface to the
findWildcardIdentities (now called addEachWildcard) which is called for
each wildcard identity of a word.