Chang wildcardMap value from HashSet<String> to HashSet<Character>
Since each word only has one edit from the wildcard identity, only the
one character difference needs to be stored. In the case of a deletion
(like the word cat and one of its identities c*at), the asterisk
represents the null character which gets deleted when converting to
string. The null character is represented by an asterisk.