Remove blacklist from findNeighbors
It was annoying to have to create an empty blacklist when one just
wanted to find all possible neighbors of a word. Additionally, it should
be equally fast (or even faster) to simply subtract the blacklist from
the neighbor set in LevenshteinGraph, as LazyDatabase had to check if a
word exists in the blacklist for each word that got added. This is
likely slower than what what was done in every other implementation of
LevenshteinDatabase and LevenshteinGraph.