Add ConnectedGraph
ConnectedGraph is a draft implementation of a BFS of the connected graph
originating from a single word. It's slow (taking multiple seconds per
word), but right now that's not the significance. It can be used to
approximate how many paths there are in total... and based on the
results from cat (5.28 million), it can be assumed that the total is in
the realm of hundreds of billions or even trillions.