Add dictionary helpers to LevenshteinDatabase
Includes the Comparator used to generate the dictionary, a method which
returns the word at a given index, and a method which searches for the
index of a given word. This is all in preparation for refactoring this
project to process words as ints instead of Strings, in order to reduce
memory usage significantly.