Refactor entire project to store words as ints/Integers internally
Since the dictionary is a static array and the fastest databases do not
need to read the String to determine if two words are adjacent (After
initialization), it saves memory and computation time to do this. Note
that nothing displays itself as a string properly, and CacheDatabase has
not yet been implemented. However, WildcardDatabase is faster than
before (and lazy appears about the same, which is expected).