Remove outerIntersects and modify outerIntersection
Having a separate outerIntersects is just boilerplate since it saves
little (if any) time in the case that they do intersect and requires
repeating work in the case they do. Also renamed getOuterIntersection,
made it return an ArrayList instead of HashSet (ArrayLists are better if
the only operation you care about is iteration) and made it static.