Ionize cesium (C# refactor)
The main goal was to convert the Java code to its *closest equivalent*
in C#, even if such equivalent does not make much sense for C# code
(avoiding the use of Properties for example). WildcardMap IO is broken
currently and the code runs twice as slowly as its Java counterpart (~70
vs ~35 microseconds monkey -> business), but otherwise works completely.