Replace wordAt and getWordIndex methods with Properties
Properties provide a cleaner way of conveying the same information, as
using a property clarifies that all you're doing is accessing
information, not producing undesired changes to the internal state or
running a computationally expensive action. Though it could be argued
that a binary search is computationally expensive, I don't believe it is
for the purposes of this.