Add MetaAnalyzer proof-of-concept
MetaAnalyzer is an attempt to represent every path in the entire
Levenshtein dictionary efficiently, both in terms of time and space. The
proof-of-concept outputs all paths connecting between a word in the
outside of the LevenshteinGraph and the root word in a format that both
generates quickly and can be parsed quickly.