Add compressed path output to LevenshteinGraph
OuterPathString and WritePathStreams both output a compressed
representation of the paths between two words, making it easier to
interpret and saving storage space. Still, both the binary and String
representations are too large to feasibly generate for the entire
dictionary, taking up over 100 megabytes for just the word "a" for both.