Add graph binary output
LevenshteinGraph's WriteOuterBinary aids in writing just the outer layer
to a binary stream. GraphToBinary writes an entire graph as a series of
these outer binary streams. Upon testing, the word "a" creates a file
size of about 434 kilobytes, which would put the entire dictionary
within the range of gigabytes, but not terabytes. This just brought the
challenge of outputting graphs describing the entire Levenshtein problem
from completely impossible to janky, but possible.