Add methods for generating graph diagnostics without finding paths
Avoiding finding the paths themselves saves a lot of memory, which will
be critical to speed up the process for finding the word pairs with the
longest and most paths. A struct was used instead of a class due to its
lower footprint and ability to be stored directly in data structures,
instead of as a pointer.