Enable multithreaded path diagnostics collection for entire dictionaries
Multithreading is necessary in order for the analysis of large
dictionaries to take an hour instead of hours. Each thread is assigned
partitions of words which are equally spaced apart in order to ensure
each thread is always working without complex scheduling (word position
is correlated with the corresponding graph size). Will eventually need
to add command-line flags to change the partitioning parameters and
more readable output.