Add timing for meta analysis
Added timing around each partition in the meta analysis. Gives us a baseline to work from before any changes at all.

I had to fix the broken threading issue: closure capture is by reference, so threads weren't getting assigned corrected ids.

Baseline perf data on my laptop, release build, on-battery.

Done with partition (2): 1 out of 100 in 41553ms
Done with partition (6): 1 out of 100 in 41926ms
Done with partition (8): 1 out of 100 in 45047ms
Done with partition (10): 1 out of 100 in 47341ms
Done with partition (0): 1 out of 100 in 48425ms
Length: 54
Count: 48
Word1: 158400
Word2: 244293
Length: 54
Count: 48
Word1: 158400
Word2: 244376
Length: 35
Count: 12204
Word1: 158400
Word2: 171272
Done with partition (11): 1 out of 100 in 49695ms
Done with partition (4): 1 out of 100 in 49824ms
Done with partition (9): 1 out of 100 in 49897ms
Done with partition (1): 1 out of 100 in 50700ms
Done with partition (3): 1 out of 100 in 53195ms
Done with partition (12): 1 out of 100 in 54869ms
Done with partition (5): 1 out of 100 in 55213ms
Done with partition (15): 1 out of 100 in 55574ms
Done with partition (14): 1 out of 100 in 55842ms
Done with partition (7): 1 out of 100 in 57768ms
Done with partition (13): 1 out of 100 in 58976ms
Done with partition (2): 2 out of 100 in 43933ms
Done with partition (6): 2 out of 100 in 50010ms
Done with partition (0): 2 out of 100 in 46426ms
Length: 54
Count: 48
Word1: 158400
Word2: 244293
Length: 54
Count: 48
Word1: 158400
Word2: 244376
Length: 23
Count: 17728
Word1: 169616
Word2: 196732
Done with partition (8): 2 out of 100 in 49973ms
Done with partition (10): 2 out of 100 in 48729ms
Done with partition (9): 2 out of 100 in 48337ms
Done with partition (12): 2 out of 100 in 44062ms
Done with partition (1): 2 out of 100 in 48382ms
Done with partition (3): 2 out of 100 in 47721ms
Done with partition (4): 2 out of 100 in 51922ms
Done with partition (5): 2 out of 100 in 46864ms
Done with partition (11): 2 out of 100 in 52697ms
Done with partition (14): 2 out of 100 in 50601ms
Done with partition (15): 2 out of 100 in 53131ms
Done with partition (13): 2 out of 100 in 50120ms
Done with partition (7): 2 out of 100 in 55489ms