Fix all warnings
Warnings frequently reveal real problems in your code. It's best to fix them all. Many teams turn on the "all warnings as errors" compiler flag, and I highly recommend this for any advanced project.

Some warnings are simple and don't matter. But a non-empty warnings list trains you to ignore all warnings, and some warnings really, really matter. So always keep your warning list empty, and don't write them off.

These were the warnings:

0>LevenshteinGraph.cs(131,13): Warning CS0472 : The result of the expression is always 'false' since a value of type 'bool' is never equal to 'null' of type 'bool?'
0>FinderSingleSided.cs(18,24): Warning CS8603 : Possible null reference return.
0>FinderSingleSided.cs(13,17): Warning CS0162 : Unreachable code detected
0>FinderDualSided.cs(25,24): Warning CS8603 : Possible null reference return.
0>FinderDualSided.cs(29,17): Warning CS0162 : Unreachable code detected
0>MetaAnalyzer.cs(5,19): Warning CS0219 : The variable 'count' is assigned but its value is never used

In this case, most were minor. The always false comparison, though, was a real performance issue in a perf-critical piece of code. Nullability is easily handled in C# and I recommend paying attention to it.

New numbers:

Done with partition (6): 1 out of 100 in 38311ms
Done with partition (2): 1 out of 100 in 38730ms
Done with partition (8): 1 out of 100 in 41874ms
Done with partition (10): 1 out of 100 in 43335ms
Done with partition (0): 1 out of 100 in 44665ms
Length: 54
Count: 48
Word1: 158400
Word2: 244293
Length: 54
Count: 48
Word1: 158400
Word2: 244376
Length: 35
Count: 12204
Word1: 158400
Word2: 171272
Done with partition (4): 1 out of 100 in 45520ms
Done with partition (11): 1 out of 100 in 45664ms
Done with partition (9): 1 out of 100 in 45679ms
Done with partition (1): 1 out of 100 in 46635ms
Done with partition (3): 1 out of 100 in 48814ms
Done with partition (12): 1 out of 100 in 50672ms
Done with partition (15): 1 out of 100 in 50809ms
Done with partition (5): 1 out of 100 in 51281ms
Done with partition (14): 1 out of 100 in 51356ms
Done with partition (7): 1 out of 100 in 52504ms
Done with partition (13): 1 out of 100 in 53309ms
Done with partition (2): 2 out of 100 in 42072ms
Done with partition (6): 2 out of 100 in 46277ms
Done with partition (0): 2 out of 100 in 42804ms
Length: 54
Count: 48
Word1: 158400
Word2: 244293
Length: 54
Count: 48
Word1: 158400
Word2: 244376
Length: 23
Count: 17728
Word1: 169616
Word2: 196732
Done with partition (8): 2 out of 100 in 46157ms
Done with partition (10): 2 out of 100 in 45376ms
Done with partition (9): 2 out of 100 in 43569ms
Done with partition (12): 2 out of 100 in 41315ms
Done with partition (1): 2 out of 100 in 45644ms
Done with partition (3): 2 out of 100 in 43851ms
Done with partition (11): 2 out of 100 in 47700ms
Done with partition (4): 2 out of 100 in 47905ms
Done with partition (5): 2 out of 100 in 43258ms
Done with partition (14): 2 out of 100 in 46924ms
Done with partition (13): 2 out of 100 in 46652ms
Done with partition (15): 2 out of 100 in 49659ms
Done with partition (7): 2 out of 100 in 50903ms

2nd partition for thread 0 46426ms -> 42804ms, so ~7.8%.