Improve path count algorithm in analyzer
The old algorithm did a lot of redundant work. So much so that just computing how many paths from the root to each node took about half the time.

This memoizes the results as we go, ensuring we only need to look one level back for each new node.

Thread 0's partition 2 time goes from 42804ms -> 28212ms, for a 34.1% win over the last improvement, and a 39.2% win overall.

There are a lot of ways to try to do this. My first try was to integrate this into the core algorithm, with a map from int to a struct of {List<int>, pathCount}. This worked, and was an improvement, but wasn't as much as I would have liked. Making those core data structures larger was a net loss vs. this way.

New numbers:

one with partition (2): 1 out of 100 in 24854ms
Done with partition (6): 1 out of 100 in 26018ms
Done with partition (10): 1 out of 100 in 27680ms
Done with partition (8): 1 out of 100 in 28145ms
Done with partition (0): 1 out of 100 in 28212ms
Length: 54
Count: 48
Word1: 158400
Word2: 244293
Length: 54
Count: 48
Word1: 158400
Word2: 244376
Length: 35
Count: 12204
Word1: 158400
Word2: 171272
Done with partition (4): 1 out of 100 in 28582ms
Done with partition (1): 1 out of 100 in 29847ms
Done with partition (9): 1 out of 100 in 30071ms
Done with partition (11): 1 out of 100 in 30117ms
Done with partition (3): 1 out of 100 in 30176ms
Done with partition (15): 1 out of 100 in 30230ms
Done with partition (14): 1 out of 100 in 30504ms
Done with partition (5): 1 out of 100 in 30624ms
Done with partition (7): 1 out of 100 in 31445ms
Done with partition (12): 1 out of 100 in 32486ms
Done with partition (13): 1 out of 100 in 34147ms
Done with partition (2): 2 out of 100 in 26764ms
Done with partition (0): 2 out of 100 in 28562ms
Length: 54
Count: 48
Word1: 158400
Word2: 244293
Length: 54
Count: 48
Word1: 158400
Word2: 244376
Length: 23
Count: 17728
Word1: 169616
Word2: 196732
Done with partition (8): 2 out of 100 in 29076ms
Done with partition (6): 2 out of 100 in 31410ms
Done with partition (10): 2 out of 100 in 30678ms
Done with partition (1): 2 out of 100 in 28767ms
Done with partition (3): 2 out of 100 in 28605ms
Done with partition (12): 2 out of 100 in 26908ms
Done with partition (4): 2 out of 100 in 30852ms
Done with partition (9): 2 out of 100 in 30374ms
Done with partition (5): 2 out of 100 in 30149ms
Done with partition (11): 2 out of 100 in 30683ms
Done with partition (7): 2 out of 100 in 29946ms
Done with partition (15): 2 out of 100 in 31746ms
Done with partition (14): 2 out of 100 in 31958ms
Done with partition (13): 2 out of 100 in 30157ms