Reduce re-allocation in new path count algorithm
This is a fun one. We know that the dictionary for the path counts is going to trend towards the total word size (or at least the size of the largest connected component), which huge. Thus the dict will have to resize and rehash many, many times. I haven't tried to figure out how many times, but it hurts.

Set an initial capacity of 300,000. It's bigger than the largest connected component, and leaves lots of room to avoid collisions.

Thread 0 partition 2 time 28562ms -> 26830ms, for a ~6.1% win.

New numbers:

Done with partition (6): 1 out of 100 in 23852ms
Done with partition (2): 1 out of 100 in 24401ms
Done with partition (10): 1 out of 100 in 25698ms
Done with partition (0): 1 out of 100 in 26972ms
Length: 54
Count: 48
Word1: 158400
Word2: 244293
Length: 54
Count: 48
Word1: 158400
Word2: 244376
Length: 35
Count: 12204
Word1: 158400
Word2: 171272
Done with partition (8): 1 out of 100 in 27732ms
Done with partition (4): 1 out of 100 in 27754ms
Done with partition (11): 1 out of 100 in 27987ms
Done with partition (1): 1 out of 100 in 28538ms
Done with partition (15): 1 out of 100 in 28733ms
Done with partition (3): 1 out of 100 in 29230ms
Done with partition (9): 1 out of 100 in 29249ms
Done with partition (7): 1 out of 100 in 29801ms
Done with partition (5): 1 out of 100 in 29956ms
Done with partition (14): 1 out of 100 in 30699ms
Done with partition (13): 1 out of 100 in 31516ms
Done with partition (12): 1 out of 100 in 31587ms
Done with partition (2): 2 out of 100 in 26402ms
Done with partition (0): 2 out of 100 in 26830ms
Length: 54
Count: 48
Word1: 158400
Word2: 244293
Length: 54
Count: 48
Word1: 158400
Word2: 244376
Length: 23
Count: 17728
Word1: 169616
Word2: 196732
Done with partition (6): 2 out of 100 in 30261ms
Done with partition (1): 2 out of 100 in 27026ms
Done with partition (10): 2 out of 100 in 29935ms
Done with partition (8): 2 out of 100 in 29219ms
Done with partition (4): 2 out of 100 in 29312ms
Done with partition (3): 2 out of 100 in 28076ms
Done with partition (12): 2 out of 100 in 26310ms
Done with partition (9): 2 out of 100 in 29316ms
Done with partition (11): 2 out of 100 in 30588ms
Done with partition (5): 2 out of 100 in 28991ms
Done with partition (15): 2 out of 100 in 31443ms
Done with partition (7): 2 out of 100 in 30871ms
Done with partition (13): 2 out of 100 in 29565ms
Done with partition (14): 2 out of 100 in 30796ms