Tuned Dict sizes in graph, removed extra hash
In aggregate these give a little boost. In theory, tuning the initial capacity of the dicts in graph ought to be a nice win. It's not as much as I would like yet. There's a wide vareity of needed sizes, so picking a single one is rough. What we really want to do is have a reusable graph which can keep the allocations and reach a steady state...

Also removed an extra hash in the inner loop.

Thread 0 partition 2 time 26830ms -> 24667ms for ~8.1%.

New numbers:

Done with partition (2): 1 out of 100 in 22356ms
Done with partition (6): 1 out of 100 in 23221ms
Done with partition (10): 1 out of 100 in 24989ms
Done with partition (4): 1 out of 100 in 25980ms
Done with partition (0): 1 out of 100 in 26014ms
Length: 54
Count: 48
Word1: 158400
Word2: 244293
Length: 54
Count: 48
Word1: 158400
Word2: 244376
Length: 35
Count: 12204
Word1: 158400
Word2: 171272
Done with partition (8): 1 out of 100 in 26032ms
Done with partition (15): 1 out of 100 in 26778ms
Done with partition (1): 1 out of 100 in 26955ms
Done with partition (3): 1 out of 100 in 27368ms
Done with partition (9): 1 out of 100 in 27378ms
Done with partition (5): 1 out of 100 in 28355ms
Done with partition (11): 1 out of 100 in 28612ms
Done with partition (7): 1 out of 100 in 28623ms
Done with partition (14): 1 out of 100 in 28814ms
Done with partition (12): 1 out of 100 in 29421ms
Done with partition (13): 1 out of 100 in 29725ms
Done with partition (2): 2 out of 100 in 24064ms
Done with partition (0): 2 out of 100 in 24667ms
Length: 54
Count: 48
Word1: 158400
Word2: 244293
Length: 54
Count: 48
Word1: 158400
Word2: 244376
Length: 23
Count: 17728
Word1: 169616
Word2: 196732
Done with partition (6): 2 out of 100 in 28212ms
Done with partition (1): 2 out of 100 in 24834ms
Done with partition (8): 2 out of 100 in 26642ms
Done with partition (10): 2 out of 100 in 27921ms
Done with partition (3): 2 out of 100 in 25556ms
Done with partition (4): 2 out of 100 in 27721ms
Done with partition (12): 2 out of 100 in 24383ms
Done with partition (15): 2 out of 100 in 27498ms
Done with partition (5): 2 out of 100 in 26190ms
Done with partition (9): 2 out of 100 in 27217ms
Done with partition (7): 2 out of 100 in 27368ms
Done with partition (11): 2 out of 100 in 28262ms
Done with partition (13): 2 out of 100 in 27559ms
Done with partition (14): 2 out of 100 in 28580ms