Apply some C# niceties
C# can be pretty terse compared to Java, which often feels more readable. I went ahead and applied all the suggestions Rider had for the graph class, just as an example.

Some of it will be personal preference. I will argue, though, in favor of var in both C# and Java. Sometimes it's obfuscating, but most often it adds wonderful flexibility as you change the code over time. Often a change of type in one place results in the corresponding type change elsewhere, and var can reduce the edits needed to make that happen. A good idea will annotate var with the real type for you, often in a smaller, somewhat unobtrusive font.