Join with analysis threads and print final results
Changed the Thread handling to give each one a place to store its results, and have the main thread join with them and aggregate the final result, and print it.

There's a bunch of ways to do this kind of thing, but this way is pretty common. You could also have a "worker" class and give one to each thread, or do crazy things having them synchronize with each other to share results, etc., but this is simple and usually pretty clear.

Removed the special case handling of the "excess" portions, too.

Modified some of the printing to be more terse.

This change is perf-neutral.

Also rubbed a little C#-ness on it.

Did this on the plane yesterday but forgot to commit!