Reuse path count memory across graphs
Hoist the pathCounts dict out of analysis of each individual word, but still per-thread. This allows us to completely reuse the dictionary memory on this one, and resetting it is very easy. Just set the root to 1, and because we BFS for every other node, and update per-level, all old data is overwritten with correct data before being read, and any extra data is simply ignored.

Thread 0, partition 2: 24667ms --> 23866ms, so ~3.2%.

New data:

Done with partition (2): 1 out of 100 in 20846ms
Done with partition (6): 1 out of 100 in 21445ms
Done with partition (10): 1 out of 100 in 23287ms
Done with partition (0): 1 out of 100 in 24058ms
Done with partition (8): 1 out of 100 in 24153ms
Done with partition (4): 1 out of 100 in 24603ms
Done with partition (11): 1 out of 100 in 25351ms
Done with partition (1): 1 out of 100 in 25594ms
Done with partition (9): 1 out of 100 in 25759ms
Done with partition (15): 1 out of 100 in 25562ms
Done with partition (3): 1 out of 100 in 26076ms
Done with partition (5): 1 out of 100 in 26129ms
Done with partition (7): 1 out of 100 in 26420ms
Done with partition (14): 1 out of 100 in 27079ms
Done with partition (12): 1 out of 100 in 27726ms
Done with partition (13): 1 out of 100 in 28886ms
Done with partition (2): 2 out of 100 in 22534ms
Done with partition (0): 2 out of 100 in 23866ms  <--
Done with partition (6): 2 out of 100 in 26835ms
Done with partition (10): 2 out of 100 in 25706ms
Done with partition (8): 2 out of 100 in 25034ms
Done with partition (1): 2 out of 100 in 24142ms
Done with partition (12): 2 out of 100 in 22255ms
Done with partition (9): 2 out of 100 in 24755ms
Done with partition (3): 2 out of 100 in 24491ms
Done with partition (4): 2 out of 100 in 26323ms
Done with partition (5): 2 out of 100 in 25025ms
Done with partition (11): 2 out of 100 in 26507ms
Done with partition (7): 2 out of 100 in 26130ms
Done with partition (15): 2 out of 100 in 26854ms
Done with partition (14): 2 out of 100 in 26267ms
Done with partition (13): 2 out of 100 in 25109ms