Reuse graphs across analyses per-thread
Create one graph for each analysis thread, and reset it for each new analysis. Currently, it only reuses the searched set, which is the largest set (and on LOH). This is a nice win.

You can probably see what's coming next, and where the limit is gonna be.

Thread 0, partition 2: 23866ms --> 20800ms, ~12.8%. We up to 46426ms --> 20800ms since we started, for ~55.2% overall.

New numbers:

one with partition (6): 1 out of 100 in 19146ms
Done with partition (2): 1 out of 100 in 19217ms
Done with partition (0): 1 out of 100 in 20511ms
Done with partition (4): 1 out of 100 in 21066ms
Done with partition (10): 1 out of 100 in 21108ms
Done with partition (8): 1 out of 100 in 21146ms
Done with partition (15): 1 out of 100 in 21340ms
Done with partition (11): 1 out of 100 in 21587ms
Done with partition (9): 1 out of 100 in 21830ms
Done with partition (3): 1 out of 100 in 21936ms
Done with partition (1): 1 out of 100 in 22054ms
Done with partition (14): 1 out of 100 in 22374ms
Done with partition (5): 1 out of 100 in 22779ms
Done with partition (7): 1 out of 100 in 22916ms
Done with partition (12): 1 out of 100 in 24476ms
Done with partition (13): 1 out of 100 in 24952ms
Done with partition (2): 2 out of 100 in 20364ms
Done with partition (0): 2 out of 100 in 20800ms <--
Done with partition (6): 2 out of 100 in 23084ms
Done with partition (3): 2 out of 100 in 21100ms
Done with partition (1): 2 out of 100 in 21062ms
Done with partition (9): 2 out of 100 in 21411ms
Done with partition (8): 2 out of 100 in 22124ms
Done with partition (4): 2 out of 100 in 22287ms
Done with partition (10): 2 out of 100 in 22286ms
Done with partition (15): 2 out of 100 in 23096ms
Done with partition (11): 2 out of 100 in 23028ms
Done with partition (12): 2 out of 100 in 20628ms
Done with partition (5): 2 out of 100 in 22592ms
Done with partition (7): 2 out of 100 in 22580ms
Done with partition (14): 2 out of 100 in 23695ms
Done with partition (13): 2 out of 100 in 21257ms