Double buffer outer set in the graph
Add some simple double buffering to the outer set in the graph. This ensures we allocate only two outer sets, and they grow to a steady state as a graph is used. Since we reuse graphs during analysis, these sets stable after a short period, and we never have to reallocate them. Resetting them requires clearing, which isn't necessarily cheap but is still quite a nice win.

Thread 0, partition 2: 20800ms --> 18133ms, ~12.8%.

Overall, we're up to 46426ms --> 18133ms, so ~60.1%.

New numbers:

Done with partition (6): 1 out of 100 in 14314ms
Done with partition (2): 1 out of 100 in 14778ms
Done with partition (10): 1 out of 100 in 15737ms
Done with partition (4): 1 out of 100 in 16074ms
Done with partition (8): 1 out of 100 in 16183ms
Done with partition (0): 1 out of 100 in 16304ms
Done with partition (9): 1 out of 100 in 16869ms
Done with partition (3): 1 out of 100 in 16869ms
Done with partition (11): 1 out of 100 in 17156ms
Done with partition (1): 1 out of 100 in 17260ms
Done with partition (5): 1 out of 100 in 17376ms
Done with partition (14): 1 out of 100 in 17543ms
Done with partition (15): 1 out of 100 in 18111ms
Done with partition (7): 1 out of 100 in 18442ms
Done with partition (12): 1 out of 100 in 18840ms
Done with partition (13): 1 out of 100 in 19117ms
Done with partition (2): 2 out of 100 in 16619ms
Done with partition (6): 2 out of 100 in 18914ms
Done with partition (0): 2 out of 100 in 18133ms <--
Done with partition (8): 2 out of 100 in 18523ms
Done with partition (10): 2 out of 100 in 19121ms
Done with partition (1): 2 out of 100 in 17786ms
Done with partition (3): 2 out of 100 in 18291ms
Done with partition (9): 2 out of 100 in 18528ms
Done with partition (4): 2 out of 100 in 19481ms
Done with partition (12): 2 out of 100 in 17028ms
Done with partition (5): 2 out of 100 in 18825ms
Done with partition (11): 2 out of 100 in 19686ms
Done with partition (13): 2 out of 100 in 19043ms
Done with partition (7): 2 out of 100 in 20014ms
Done with partition (14): 2 out of 100 in 20914ms
Done with partition (15): 2 out of 100 in 21680ms