Set initial capacity on parent lists in graph
The default initial size on List<int> is smaller than our typical use of it as a parent list in the graph. I've set the initial to 16, which reduces reallocation greatly. I didn't try to find the best number, I tried 8, 16, and 32 and liked 16 best.

Thread 0, partition 2, 18133ms --> 10046ms, ~44.6%. Wow.

Overall, 46426ms --> 10046ms, ~78.4%

New numbers:

Done with partition (2): 1 out of 100 in 8085ms
Done with partition (6): 1 out of 100 in 8448ms
Done with partition (10): 1 out of 100 in 9109ms
Done with partition (0): 1 out of 100 in 9375ms
Done with partition (8): 1 out of 100 in 9522ms
Done with partition (4): 1 out of 100 in 9710ms
Done with partition (1): 1 out of 100 in 9918ms
Done with partition (11): 1 out of 100 in 9939ms
Done with partition (3): 1 out of 100 in 10053ms
Done with partition (5): 1 out of 100 in 10139ms
Done with partition (9): 1 out of 100 in 10155ms
Done with partition (14): 1 out of 100 in 10197ms
Done with partition (15): 1 out of 100 in 10392ms
Done with partition (7): 1 out of 100 in 10528ms
Done with partition (12): 1 out of 100 in 10852ms
Done with partition (13): 1 out of 100 in 11031ms
Done with partition (2): 2 out of 100 in 9136ms
Done with partition (6): 2 out of 100 in 10720ms
Done with partition (0): 2 out of 100 in 10046ms <--
Done with partition (8): 2 out of 100 in 10344ms
Done with partition (1): 2 out of 100 in 9956ms
Done with partition (10): 2 out of 100 in 10894ms
Done with partition (3): 2 out of 100 in 10220ms
Done with partition (12): 2 out of 100 in 9657ms
Done with partition (4): 2 out of 100 in 10829ms
Done with partition (9): 2 out of 100 in 10456ms
Done with partition (5): 2 out of 100 in 10719ms
Done with partition (11): 2 out of 100 in 11222ms
Done with partition (14): 2 out of 100 in 11429ms
Done with partition (7): 2 out of 100 in 11102ms
Done with partition (13): 2 out of 100 in 11015ms
Done with partition (15): 2 out of 100 in 11996ms