Remove graph binary stream output
Binary stream output was added to test whether outputting graphs to a
file was viable. Due to the large file sizes involved, it does not seem
like it will ever be viable. Removing rather than deprecating since
backwards compatibility is not a concern for a project with no official
release.