Remove searched count and silence unreachable code warnings
Printing the size of searched is redundant, as it is simply the sum of
the previous outer sizes. The unreachable code compiler warnings are
unhelpful since the entire point is for the if statement to only be true
if the programmer changes the PRINT_EXTRA constant. Though this is
likely poor design, changing it is not currently a priority.