Add HashSetBFSGraph
This is backed by a List of HashSets instead of two dictionaries.
Instead of storing a new list for the previous words from each word, it
compares its neighbors to the words contained in layers of the graph,
saving a lot of memory. Its anywhere from twice as fast (dog to cat) to
sligtly slower (dog to quack) than DictionaryBFSGraph. Nothing uses it
yet due to its reliance on CacheDatabase.